fix: Moving more tests that have hidden dependencies on AWS or sops
