+++ test/integration_test.go
@@ -3373,25 +3373,6 @@ func TestDependencyOutputModulePrefix(t *testing.T) {
 	assert.Equal(t, 42, int(outputs["z"].Value.(float64)))
 }
 
-func TestErrorExplaining(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := helpers.CopyEnvironment(t, testFixtureInitError)
-	initTestCase := util.JoinPath(tmpEnvPath, testFixtureInitError)
-
-	helpers.CleanupTerraformFolder(t, initTestCase)
-	helpers.CleanupTerragruntFolder(t, initTestCase)
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := helpers.RunTerragruntCommand(t, "terragrunt init -no-color --terragrunt-forward-tf-stdout --terragrunt-non-interactive --terragrunt-working-dir "+initTestCase, &stdout, &stderr)
-	require.Error(t, err)
-
-	explanation := shell.ExplainError(err)
-	assert.Contains(t, explanation, "Check your credentials and permissions")
-}
-
 func TestExplainingMissingCredentials(t *testing.T) {
 	// no parallel because we need to set env vars
 	t.Setenv("AWS_SHARED_CREDENTIALS_FILE", "/tmp/not-existing-creds-46521694")
@@ -3717,22 +3698,6 @@ func TestTerragruntHandleEmptyStateFile(t *testing.T) {
 	helpers.CreateEmptyStateFile(t, testPath)
 
 	helpers.RunTerragrunt(t, "terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+testPath)
-}
-
-func TestTerragruntInvokeTerraformTests(t *testing.T) {
-	t.Parallel()
-	if isTerraform() {
-		t.Skip("Not compatible with Terraform 1.5.x")
-		return
-	}
-
-	tmpEnvPath := helpers.CopyEnvironment(t, testFixtureTfTest)
-	helpers.CleanupTerraformFolder(t, tmpEnvPath)
-	testPath := util.JoinPath(tmpEnvPath, testFixtureTfTest)
-
-	stdout, _, err := helpers.RunTerragruntCommandWithOutput(t, "terragrunt test --terragrunt-non-interactive --terragrunt-forward-tf-stdout --terragrunt-working-dir "+testPath)
-	require.NoError(t, err)
-	assert.Contains(t, stdout, "1 passed, 0 failed")
 }
 
 func TestTerragruntCommandsThatNeedInput(t *testing.T) {
